add the-champion-gets-more-children strategy.
After 5-10 generations, improvements tend to happen only for the
best individual.

In this strategy, when the champion keeps winning the competition for n times,
the champion can produce n children.

On the other hands, the number of "winner"s of the competition shrinks accordingly,
so that we can keep the same population size.